Gather the join's output columns by position (#895)
The probe loop built a Vec<Value> for the driving row, another for every
match, pushed them into a list and transposed the list back into columns
at the end of the chunk. None of that was the join. The answer is the
driving side gathered by one list of positions and the gathered side
gathered by another, so the probe now writes the two lists and gathers
once per column.

Adds side.rs, where Build holds the gathered side as columns laid end to
end by Assembly. PAD is u32::MAX and a gather past the end already reads
as null, so an unmatched row on a left join is a position like any other
and the gather has no branch in it.

The build side takes the Keep sink rather than Gather, so it is no longer
taken apart into a row per row on the way in. Lookup::matches yields u32
because what the caller does with it is gather. The driving row is only
boxed into values when a residual conjunct needs it, so a plain equi join
allocates no row at all.

Ten million probe rows against a one million row build on gpc: 0.110s
wall, 0.449s CPU, 45ns per probe row, against 0.347s and 232ns before.
DuckDB is 0.037s and 27ns on the same shape.

Closes #880.

+++ crates/rudb-exec/src/lookup.rs
@@ -256,14 +256,18 @@ impl Lookup {
     ///
     /// `into` is the caller's buffer so that a driving row does not cost an allocation, and it is
     /// cleared here rather than by the caller.
-    pub(crate) fn matches(&self, slot: usize, into: &mut Vec<usize>) {
+    ///
+    /// Row numbers rather than indices, because what the caller does with them is hand them to
+    /// [`Build::gather`](crate::side::Build::gather), and a gather takes a run of `u32`. The chain
+    /// is a run of `u32` already, so this is a copy rather than a widening.
+    pub(crate) fn matches(&self, slot: usize, into: &mut Vec<u32>) {
         into.clear();
         if slot == MISS {
             return;
         }
         let mut at = self.head[slot];
         while at != NONE {
-            into.push(at as usize);
+            into.push(at);
             at = self.next[at as usize];
         }
     }
@@ -339,7 +343,7 @@ mod tests {
     }
 
     /// What one driving row of the same shape finds, in the order it finds it.
-    fn found(lookup: &Lookup, values: &[Option<i32>]) -> Vec<Vec<usize>> {
+    fn found(lookup: &Lookup, values: &[Option<i32>]) -> Vec<Vec<u32>> {
         let mut scratch = Scratch::default();
         let mut slots = Vec::new();
         lookup.slots(&[column(values)], values.len(), &[false], &mut scratch, &mut slots);
@@ -390,7 +394,7 @@ mod tests {
         let mut slots = Vec::new();
         lookup.slots(&[column(&[Some(1)])], 1, &[false], &mut scratch, &mut slots);
         lookup.matches(slots[0], &mut chain);
-        let wanted: Vec<usize> = (0..500).filter(|row| row % 3 == 1).collect();
+        let wanted: Vec<u32> = (0..500).filter(|row| row % 3 == 1).collect();
         assert_eq!(chain, wanted);
     }
